+++ modules/@apostrophecms/doc-type/ui/apos/components/AposDocLocalePicker.vue
@@ -16,11 +16,11 @@
     >
       <AposLocalePicker
         :current-locale="locale"
-        :localized="localized"
+        :localized="docLocalized"
         :forbidden="forbidden"
         :forbidden-tooltip="forbiddenTooltip"
         :is-open="isOpen"
-        :show-localized="showLocalized"
+        :show-localized="!isManager"
         @switch-locale="switchLocale"
       />
     </AposContextMenu>
@@ -53,19 +53,18 @@ const props = defineProps({
     type: Boolean,
     required: true
   },
-  showLocalized: {
+  isManager: {
     type: Boolean,
-    default: true
+    default: false
   }
 });
 
 const $t = inject('i18n');
 const i18nAction = apos.modules['@apostrophecms/i18n'].action;
-const forbiddenTooltip = $t('apostrophe:localeSwitcherPermissionToCreate', {
-  docType: props.moduleOptions.label.toLowerCase()
-});
+const docType = $t(props.moduleOptions.label)?.toLowerCase();
+const forbiddenTooltip = $t('apostrophe:localeSwitcherPermissionToCreate', { docType });
 const menu = ref(null);
-const localized = ref({});
+const docLocalized = ref({});
 const forbidden = ref([]);
 const isOpen = ref(false);
 const button = computed(() => {
@@ -91,7 +90,7 @@ async function open() {
     const docs = await apos.http.get(
     `${props.moduleOptions.action}/${props.docId}/locales`, { busy: true }
     );
-    localized.value = Object.fromEntries(
+    docLocalized.value = Object.fromEntries(
       docs.results
         .filter(doc => doc.aposLocale.endsWith(':draft'))
         .map(doc => [ doc.aposLocale.split(':')[0], doc ])
@@ -106,7 +105,7 @@ async function open() {
 
 async function checkCreatePermission() {
   const localesWithNoDocs = Object.keys(apos.i18n.locales)
-    .filter((locale) => !localized.value[locale]);
+    .filter((locale) => !docLocalized.value[locale]);
 
   const allowed = await apos.http.get(`${i18nAction}/locales-permissions`, {
     qs: {
@@ -131,19 +130,24 @@ async function switchLocale(locale) {
     return;
   }
 
+  if (props.isManager) {
+    emit('switch-locale', { locale });
+    return;
+  }
+
   const save = props.isModified
     ? await apos.confirm({
       heading: 'apostrophe:unsavedChanges',
-      description: $t(
-        'apostrophe:localeSwitcherDiscardChangesPrompt',
-        { docType: props.moduleOptions.label.toLowerCase() }
-      ),
+      description: 'apostrophe:localeSwitcherDiscardChangesPrompt',
       negativeLabel: 'apostrophe:localeSwitcherDiscardChangesNegative',
       affirmativeLabel: 'apostrophe:localeSwitcherDiscardChangesAffirmative',
       icon: false
     }, {
       hasCloseButton: true,
-      tiny: true
+      tiny: true,
+      interpolate: {
+        docType
+      }
     })
     : false;
 
@@ -151,9 +155,33 @@ async function switchLocale(locale) {
     return;
   }
 
+  const docExist = docLocalized.value[props.locale];
+  const isDocLocalized = docLocalized.value[locale.name];
+  const toLocalize = (docExist && !isDocLocalized)
+    ? await apos.confirm({
+      heading: 'apostrophe:switchLocalesAndLocalizeDoc',
+      description: 'apostrophe:notInLocaleDoc',
+      negativeLabel: 'apostrophe:noCreateBlankDoc',
+      affirmativeLabel: 'apostrophe:yesLocalizeAndSwitchLocalesDoc'
+    }, {
+      hasCloseButton: true,
+      tiny: true,
+      interpolate: {
+        docType,
+        label: locale.label,
+        currentLocale: props.locale
+      }
+    })
+    : false;
+
+  if (toLocalize === null) {
+    return;
+  }
+
   emit('switch-locale', {
-    locale: locale.name,
-    localized: localized.value[locale.name],
+    locale,
+    localized: docLocalized.value[locale.name],
+    toLocalize,
     save
   });
 }